use short

+++ common/src/main/java/us/myles/ViaVersion/protocols/protocol1_9to1_8/packets/WorldPackets.java
@@ -292,7 +292,7 @@ public class WorldPackets {
                         // First set this packet ID to Block placement
                         wrapper.setId(0x08);
                         wrapper.write(Type.POSITION, new Position((long) -1, (long) -1, (long) -1));
-                        wrapper.write(Type.UNSIGNED_BYTE, (byte) 255);
+                        wrapper.write(Type.UNSIGNED_BYTE, (short) 255);
                         // Write item in hand
                         Item item = Protocol1_9To1_8.getHandItem(wrapper.user());
                         // Blocking patch
